requirements.txt
python-dotenv
discord.py[voice]
riotwatcher
